refactor(web): extract session management and PDF download into dedicated modules
Move CSRF/session token handling from client.ts into api/session.ts and
isolate downloadCompensationPDF into its own file for cleaner module boundaries.
