+++ .changeset/extract-session-module.md
@@ -1,0 +1,5 @@
+---
+'volleykit-web': patch
+---
+
+Extract session management into dedicated module and isolate PDF download
